requirements.txt
aiogram
psycopg2-binary
bcrypt
python-dotenv
bcrypt
